requirements.txt
anyio==3.7.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
exceptiongroup==1.1.1
fastapi==0.97.0
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.2
httpie==3.2.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
install==1.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.9
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.198.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
